Move date, time & retry utility code to `util/` dir
As per #648, move several modules containing utility functions which
have no dependencies on other code to `src/sidebar/util`.
