+++ src/sidebar/util/time.js
